separate out KvPairs
